shaco v0.2

+++ Common/shaco_stdlib.h
@@ -1,7 +1,8 @@
 #ifndef _S_STDLIB_H
 #define _S_STDLIB_H
-void *shaco_malloc(unsigned long len);
-void *shaco_calloc(unsigned long nmemb, unsigned long size);
-void *shaco_realloc(void * ptr, unsigned long size);
+#include <stddef.h>
+void *shaco_malloc(size_t len);
+void *shaco_calloc(size_t nmemb, size_t size);
+void *shaco_realloc(void * ptr, size_t size);
 void shaco_free(void *);
 #endif
